requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
python-dotenv
itsdangerous
